moved to double to increase range, started async computation experiment

+++ PrimeFactorizer/PrimeCheck.h
@@ -9,7 +9,7 @@ class PrimeCheck
 public:
 	PrimeCheck() noexcept; 
 	PrimeCheck(shared_ptr <PrimeLibrary> l);
-	bool is_prime(unsigned long long input);
+	bool is_prime(double input);
 
 private:
 	shared_ptr <PrimeLibrary> Lib;
